% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{changeFeatID}
\alias{changeFeatID}
\title{Change a feature ID symbol for another}
\usage{
changeFeatID(x, old_ID, new_ID)
}
\arguments{
\item{x}{prbList}

\item{old_ID}{vector of feature IDs to change}

\item{new_ID}{vector of feature IDs to change to}
}
\value{
prbList
}
\description{
Change a feature ID symbol for another
}
