% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{createPrbList}
\alias{createPrbList}
\title{Create a prbList object}
\usage{
createPrbList(
  probe_data,
  GO_cols,
  comm_cols = NULL,
  GO_hierarchy = NULL,
  read_separator = ":",
  obj_separator = ":"
)
}
\arguments{
\item{probe_data}{dataframe-like object containing probe information.
See details}

\item{GO_cols}{vector of column(s) in probe_data that contain user-
defined terms that organize the information.}

\item{comm_cols}{(optional) vector of column(s) to use for the comments info.
If left as NULL, all columns other than GO and feat_ID will be used.}

\item{GO_hierarchy}{(optional) user defined hierarchy can be passed here.
Should be provided as a named list of vectors, where the list names are the
parent terms and the items in the vectors are the child terms}

\item{read_separator}{(default = ":") separator to use on data to be read in}

\item{obj_separator}{(optional. default = ":") separator symbol to use in the
final prbList object}
}
\description{
Create a prbList object
}
