% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{feature}
\alias{feature}
\title{Examine feature}
\usage{
feature(x, feat, info = c("GO", "comments"))
}
\arguments{
\item{x}{prbList object}

\item{feat}{feature to examine}

\item{info}{annotation columns to pull from}
}
\description{
Examine feature
}
