% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{getFeats}
\alias{getFeats}
\title{Get features by annotation}
\usage{
\S4method{getFeats}{prbList}(x, query = NULL, by = "GO", to_clip = FALSE, panel_only = TRUE)
}
\arguments{
\item{x}{prbList object}

\item{query}{query term to search annotations by}

\item{by}{column to query by}

\item{to_clip}{write results to clipboard}

\item{in_panel}{only get features set as in panel}
}
\description{
Get features by annotation
}
