% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getGOHierarchical}
\alias{getGOHierarchical}
\title{Get GO terms in a hierarchical manner}
\usage{
getGOHierarchical(x, query, visited_terms = c())
}
\arguments{
\item{x}{prbList}

\item{query}{GO term to search for child terms with}

\item{visited_terms}{do not use. Used in recursion}
}
\description{
Get GO terms in a hierarchical manner
}
