% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{GOHierarchy-generic}
\alias{GOHierarchy-generic}
\alias{GOHierarchy,}
\alias{GOHierarchy<-}
\alias{GOHierarchy,prbList,character-method}
\alias{GOHierarchy,prbList,missing-method}
\alias{GOHierarchy<-,prbList,missing-method}
\alias{GOHierarchy<-,prbList,character-method}
\title{Gene ontology hierarchy}
\usage{
\S4method{GOHierarchy}{prbList,character}(x, parent)

\S4method{GOHierarchy}{prbList,missing}(x)

\S4method{GOHierarchy}{prbList,missing}(x) <- value

\S4method{GOHierarchy}{prbList,character}(x, parent) <- value
}
\arguments{
\item{x}{prbList object}

\item{parent}{parent GO term}

\item{value}{children GO terms to set}
}
\description{
Get and set gene ontology hierarchy
}
