% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{lockFeature}
\alias{lockFeature}
\alias{lockFeat}
\alias{unlockFeat}
\title{Feature locking}
\usage{
lockFeat(x, feats)

unlockFeat(x, feats)
}
\arguments{
\item{x}{prbList object}

\item{feats}{features to lock or unlock}
}
\description{
Lock important features to prevent easily excluding them from the
panel by accident.
}
\section{Functions}{
\itemize{
\item \code{lockFeat()}: Lock the designated features

\item \code{unlockFeat()}: Unlock the designated features

}}
