% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{panelExclude}
\alias{panelExclude}
\title{Set features as excluded}
\usage{
panelExclude(x, feats)
}
\arguments{
\item{x}{prbList object}

\item{feats}{features to exclude}
}
\description{
Exclude features from final panel but do not remove entry
}
