% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{panelInclude}
\alias{panelInclude}
\title{Set features as included}
\usage{
panelInclude(x, feats)
}
\arguments{
\item{x}{prbList object}

\item{feats}{features to include}
}
\description{
Include features in final panel.
}
