% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{prbList-class}
\alias{prbList-class}
\alias{prbList}
\title{Container for probelist information}
\description{
Container for probelist information
}
\section{Slots}{

\describe{
\item{\code{infoDT}}{data.table of gene information and annotations}

\item{\code{GO_terms}}{annotation terms}

\item{\code{GO_hierarchy}}{hierarchy information between annotation terms}

\item{\code{igraph}}{igraph of hierarchy}

\item{\code{separator}}{separator symbol used within object}
}}

