% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readClipTbl}
\alias{readClipTbl}
\title{Read clipboard as table}
\usage{
readClipTbl(header = TRUE, sep = "\\t", ...)
}
\arguments{
\item{header}{whether to read first line as colnames}

\item{sep}{separator symbol to use when reading}

\item{...}{additional params to pass to read.table()}
}
\description{
Read clipboard as table
}
