% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probeList.R
\name{setFeatID}
\alias{setFeatID}
\title{Change a feature ID symbol for another}
\usage{
setFeatID(x, old_ID, new_ID)
}
\arguments{
\item{x}{prbList}

\item{old_ID}{feature ID to change}

\item{new_ID}{feature ID to change to}
}
\value{
prbList
}
\description{
Change a feature ID symbol for another
}
