% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{validate_gene_names}
\alias{validate_gene_names}
\title{Test for non-official gene names}
\usage{
validate_gene_names(x, species = "Hs")
}
\arguments{
\item{x}{a prbList}

\item{species}{species to draw gene reference from}
}
\description{
Test feature IDs to see if they are the official gene names using
limma::alias2Symbol. If they are, TRUE is returned. Otherwise, elements that
are only in the non-corrected list and those only in the corrected list will
be printed.
}
